requirements.txt
fastapi
uvicorn
python-multipart
uuid
prisma
python-dotenv
langchain
langchain-cohere
pypdf
chromadb
langchain-chroma
ddgs
beautifulsoup4
requests
